+++ roc/flags.go
@@ -1,0 +1,6 @@
+package roc
+
+/*
+#cgo LDFLAGS: -lroc
+*/
+import "C"
