+++ roc/sender.go
@@ -1,0 +1,9 @@
+package roc
+
+/*
+#include <roc/sender.h>
+*/
+import "C"
+
+// Sender as declared in roc/sender.h:96
+type Sender C.roc_sender
